+++ finagle-netty4/src/main/scala/com/twitter/finagle/netty4/channel/ChannelStatsHandler.scala
@@ -5,12 +5,21 @@ import com.twitter.finagle.netty4.channel.ChannelStatsHandler.SharedChannelStats
 import com.twitter.finagle.stats.{StatsReceiver, Verbosity}
 import com.twitter.util.{Duration, Monitor, Stopwatch}
 import io.netty.buffer.ByteBuf
-import io.netty.channel.{ChannelDuplexHandler, ChannelHandlerContext, ChannelPromise}
+import io.netty.channel.epoll.{EpollSocketChannel, EpollTcpInfo}
+import io.netty.channel.{
+  ChannelDuplexHandler,
+  ChannelException,
+  ChannelHandlerContext,
+  ChannelPromise
+}
 import io.netty.handler.ssl.SslHandshakeCompletionEvent
 import io.netty.handler.timeout.TimeoutException
+import io.netty.util.concurrent.ScheduledFuture
 import java.io.IOException
+import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.LongAdder
 import java.util.logging.{Level, Logger}
+import scala.util.control.NonFatal
 
 private object ChannelStatsHandler {
   private val log = Logger.getLogger(getClass.getName)
@@ -40,17 +49,78 @@ private object ChannelStatsHandler {
       statsReceiver.counter(Verbosity.Debug, "socket_writable_ms")
     val unwritable =
       statsReceiver.counter(Verbosity.Debug, "socket_unwritable_ms")
+    val retransmits =
+      statsReceiver.counter(Verbosity.Debug, name = "tcp_retransmits")
+    val tcpSendWindowSize =
+      statsReceiver.stat(Verbosity.Debug, name = "tcp_send_window_size")
 
     val receivedBytes = statsReceiver.counter("received_bytes")
     val sentBytes = statsReceiver.counter("sent_bytes")
     val exceptions = statsReceiver.scope("exn")
     val closesCount = statsReceiver.counter("closes")
+
     private val connections = statsReceiver.addGauge("connections") {
       connectionCount.sum()
     }
     private val tlsConnections = statsReceiver.addGauge("tls", "connections") {
       tlsConnectionCount.sum()
     }
+  }
+
+  /**
+   * How often to update TCP stats
+   */
+  private final val TcpStatsUpdateInterval = Duration.fromSeconds(30)
+
+  /**
+   * A Runnable that continuously schedules itself for execution every `TcpStatsUpdateInterval`.
+   * This holds a reference to the channel it monitors and must be cancelled when the channel is
+   * closed.
+   *
+   * This implementation here relies on the fact that a different [[ChannelStatsHandler]] instance
+   * exists for each [[Channel]].
+   */
+  private final class TcpStatsUpdater(
+    sharedChannelStats: SharedChannelStats,
+    channel: EpollSocketChannel
+  ) extends Runnable {
+    private[this] val tcpInfo: EpollTcpInfo = new EpollTcpInfo
+    private[this] var lastRetransmits: Long = 0
+    private[this] var cancelled = false
+    private[this] val scheduledUpdate: ScheduledFuture[_] =
+      // Schedule ourself to be run in the event loop for this channel.  The epoll docs are unclear
+      // on if tcpInfo needs to be invoked from the event loop, so we err on the side of caution.
+      channel
+        .eventLoop().scheduleAtFixedRate(
+          /* runnable */ this,
+          /* initialDelay */ 0,
+          /* period */ TcpStatsUpdateInterval.inMilliseconds,
+          /* timeUnit */ TimeUnit.MILLISECONDS
+        )
+
+    def cancel(): Unit = {
+      cancelled = true
+      scheduledUpdate.cancel(false)
+    }
+
+    override def run(): Unit = if (!cancelled) updateSocketStats()
+
+    private def updateSocketStats(): Unit =
+      try {
+        channel.tcpInfo(tcpInfo)
+        // linux reports `snd_cwnd` as number of segments, so we need to multiply by
+        // `snd_mss` in order to convert to number of bytes.
+        sharedChannelStats.tcpSendWindowSize.add(tcpInfo.sndCwnd() * tcpInfo.sndMss())
+        val retransmits = tcpInfo.totalRetrans()
+        sharedChannelStats.retransmits.incr(retransmits - lastRetransmits)
+        lastRetransmits = retransmits
+      } catch {
+        case _: ChannelException =>
+        // safe to ignore, this can be thrown if the socket was closed at some point before we call
+        // tcpInfo().
+        case NonFatal(t) =>
+          log.log(Level.WARNING, "Error updating TCP info stats", t)
+      }
   }
 }
 
@@ -75,6 +145,7 @@ private class ChannelStatsHandler(sharedChannelStats: SharedChannelStats)
   private[this] var connectionDuration: Stopwatch.Elapsed = _
   private[this] var channelActive: Boolean = false
   private[this] var tlsChannelActive: Boolean = false
+  private[this] var tcpStatsUpdater: TcpStatsUpdater = _
 
   override def handlerAdded(ctx: ChannelHandlerContext): Unit = {
     channelBytesRead = 0L
@@ -88,6 +159,12 @@ private class ChannelStatsHandler(sharedChannelStats: SharedChannelStats)
     sharedChannelStats.connects.incr()
     sharedChannelStats.connectionCountIncrement()
 
+    ctx.channel() match {
+      case epsc: EpollSocketChannel =>
+        if (tcpStatsUpdater != null) tcpStatsUpdater.cancel()
+        tcpStatsUpdater = new TcpStatsUpdater(sharedChannelStats, epsc)
+      case _ =>
+    }
     channelActive = true
     connectionDuration = Stopwatch.start()
     super.channelActive(ctx)
@@ -137,6 +214,10 @@ private class ChannelStatsHandler(sharedChannelStats: SharedChannelStats)
       val oldChannelBytesWritten = channelBytesWritten
       channelBytesRead = 0
       channelBytesWritten = 0
+      if (tcpStatsUpdater != null) {
+        tcpStatsUpdater.cancel()
+        tcpStatsUpdater = null
+      }
       sharedChannelStats.connectionReceivedBytes.add(oldChannelBytesRead)
       sharedChannelStats.connectionSentBytes.add(oldChannelBytesWritten)
 
@@ -170,7 +251,6 @@ private class ChannelStatsHandler(sharedChannelStats: SharedChannelStats)
       val stat =
         if (channelWasWritable) sharedChannelStats.writable else sharedChannelStats.unwritable
       stat.incr(elapsed.inMilliseconds.toInt)
-
       channelWasWritable = isWritable
       channelWritableDuration = Stopwatch.start()
     }